pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dinov3-vit-h"
version = "0.1.0"
description = "A minimal environment for DINOv3 image feature extraction and retrieval using Hugging Face Transformers."
authors = [
    { name = "Yilin Yang", email = "yilin.yang@example.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10,<3.12"

dependencies = [
    "torch==2.3.1",
    "torchvision==0.18.1",
    "transformers==4.56.0",
    "huggingface_hub==0.34.0",
    "accelerate==0.33.0",
    "tqdm==4.66.4",
    "numpy==1.26.4",
    "pillow==10.4.0",
    "nvidia-ml-py==12.535.133"
]


